[refactor] Refactor orphan block management related functions into a single file
