+++ chain/chain/src/block_processing_utils.rs
@@ -1,5 +1,6 @@
-use crate::chain::{BlockMissingChunks, OrphanMissingChunks};
+use crate::chain::BlockMissingChunks;
 use crate::near_chain_primitives::error::BlockKnownError::KnownInProcessing;
+use crate::orphan::OrphanMissingChunks;
 use crate::Provenance;
 use near_primitives::block::Block;
 use near_primitives::challenge::{ChallengeBody, ChallengesResult};
